backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4
docling>=2.80.0

# Vector Database
qdrant-client==1.7.3

# Embeddings and AI
openai==1.12.0
google-genai>=1.0.0
sentence-transformers==2.3.1

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
numpy>=1.26.0
pandas>=2.2.0
tiktoken==0.5.2

# CORS and middleware
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
loguru==0.7.2

